+++ mayan/apps/common/warnings.py
@@ -1,0 +1,7 @@
+from __future__ import absolute_import
+
+
+class DatabaseWarning(UserWarning):
+    """
+    Warning when using unsupported database backends
+    """